RobotMovements
A robot is located in the upper-left corner of a 4×4 grid. The robot can
move either up, down, left, or right, but cannot go to the same location
twice. The robot is trying to reach the lower-right corner of the grid.
Your task is to find out the number of unique ways to reach the
destination.